Refactor: Aspect에서 파라미터 추출 로직을 이름에 의존하지 않고 타입에 의존하여 추출
