+++ server/src/main/java/com/ryc/api/v2/club/business/ClubService.java
@@ -15,7 +15,7 @@ import com.ryc.api.v2.club.presentation.dto.response.ClubCreateResponse;
 import com.ryc.api.v2.club.presentation.dto.response.ClubGetResponse;
 import com.ryc.api.v2.club.presentation.dto.response.ClubUpdateResponse;
 import com.ryc.api.v2.common.aop.annotation.HasRole;
-import com.ryc.api.v2.common.dto.ClubRoleSecuredDto;
+import com.ryc.api.v2.common.aop.dto.ClubRoleSecuredDto;
 import com.ryc.api.v2.common.exception.code.ClubErrorCode;
 import com.ryc.api.v2.common.exception.custom.ClubException;
 import com.ryc.api.v2.role.business.RoleService;
@@ -51,10 +51,10 @@ public class ClubService {
 
   @Transactional
   @HasRole(Role.MEMBER)
-  public ClubUpdateResponse updateClub(ClubRoleSecuredDto roleDto, ClubUpdateRequest body) {
+  public ClubUpdateResponse updateClub(ClubRoleSecuredDto clubRoleSecuredDto, ClubUpdateRequest body) {
     Club previousClub =
         clubRepository
-            .findById(roleDto.clubId())
+            .findById(clubRoleSecuredDto.clubId())
             .orElseThrow(() -> new ClubException(ClubErrorCode.CLUB_NOT_FOUND));
 
     if (body.name() != null && clubRepository.existsByName(body.name())) {
